# app-books/src/main/java/com/distribuida/dtos/AuthorDto.java
package com.distribuida.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AuthorDto {
    private Integer id;
    private String first_name;
    private String last_name;

}